fix(agent): move title helpers to src/shared so hand_off works in agents
`coordination_tool._hand_off_v2` imported `LONG_TITLE_THRESHOLD` and
`_normalize_title_and_description` from `src.host.orchestration`. Agent
containers ship only `src/agent` + `src/shared` (Dockerfile.agent), so
that import raised `ModuleNotFoundError: No module named 'src.host'`
the moment orchestration v2 was enabled — every `hand_off` call failed
100%.

Move the title-length policy (`MAX_TITLE_CHARS`, `LONG_TITLE_THRESHOLD`,
`SHORT_TITLE_TARGET`, `_derive_short_title`, `_normalize_title_and_description`)
into a new `src/shared/task_titles.py`. Both the host-side `Tasks.create`
and the agent-side `_hand_off_v2` import from there — no cross-zone
reach, no re-export shim. Behavior is unchanged; tests updated to point
at the canonical location.
