+++ src/host/orchestration.py
@@ -32,6 +32,7 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import Any
 
+from src.shared.task_titles import _normalize_title_and_description
 from src.shared.utils import setup_logging
 
 logger = setup_logging("host.orchestration")
@@ -57,96 +58,10 @@ VALID_OUTCOMES: frozenset[str] = frozenset({"accepted", "rework", "rejected"})
 # and the UI textarea doesn't smuggle a multi-MB blob into the table.
 MAX_FEEDBACK_CHARS: int = 2000
 
-# Title length policy. ``MAX_TITLE_CHARS`` is a hard cap — anything
-# beyond is truncated server-side. ``LONG_TITLE_THRESHOLD`` is the soft
-# ceiling: when a caller submits a single ``title`` longer than this and
-# no separate ``description``, we treat the whole string as the
-# description and derive a short title from its first sentence /
-# leading clause. This is defense-in-depth against agents that put full
-# instructions into the title field — the dashboard expects a one-line
-# label, not a paragraph.
-MAX_TITLE_CHARS: int = 200
-LONG_TITLE_THRESHOLD: int = 100
-SHORT_TITLE_TARGET: int = 80
-
-
-def _derive_short_title(text: str) -> str:
-    """Extract a short single-line label from a longer task instruction.
-
-    Strategy: take the first non-empty line, then split on sentence
-    terminators (``.``, ``!``, ``?``) and the en/em dash separators we
-    see in handoff phrasings ("Draft Q3 — full brief..."). Pick the
-    first chunk; if it's still too long, hard-cut at ``SHORT_TITLE_TARGET``
-    with an ellipsis. Whitespace-collapsed so multi-line inputs don't
-    leave dangling indents.
-
-    Returns ``""`` for empty or whitespace-only input — callers (the
-    title-normalizer / ``Tasks.create``) handle that fallback.
-    """
-    if not text or not text.strip():
-        return ""
-    # First non-empty line — handles "Subject\nBody" payloads cleanly.
-    first_line = ""
-    for raw in text.splitlines():
-        stripped = raw.strip()
-        if stripped:
-            first_line = stripped
-            break
-    if not first_line:
-        first_line = text.strip()
-    # Split on sentence + dash boundaries; pick the first chunk that
-    # actually carries content (skip empty leading splits).
-    candidate = first_line
-    for sep in (". ", "! ", "? ", " — ", " - ", ": "):
-        if sep in candidate:
-            head = candidate.split(sep, 1)[0].strip()
-            if head:
-                candidate = head
-                break
-    # Collapse internal whitespace so a copy-pasted blob doesn't render
-    # ragged in a one-line truncate.
-    candidate = " ".join(candidate.split())
-    if len(candidate) > SHORT_TITLE_TARGET:
-        # Prefer cutting on a word boundary near the cap so we don't end
-        # mid-word. ``rsplit`` returns the original string if no space.
-        cut = candidate[:SHORT_TITLE_TARGET].rsplit(" ", 1)[0]
-        if len(cut) < SHORT_TITLE_TARGET // 2:
-            cut = candidate[:SHORT_TITLE_TARGET]
-        candidate = cut.rstrip(",;:") + "…"
-    return candidate
-
-
-def _normalize_title_and_description(
-    title: str, description: str | None,
-) -> tuple[str, str | None]:
-    """Apply the title-length policy.
-
-    Three cases:
-
-    * Title is short (≤ ``LONG_TITLE_THRESHOLD``): pass through unchanged.
-    * Title is long but caller provided a separate ``description``: trust
-      the caller (they made an intentional choice), just hard-cap title
-      at ``MAX_TITLE_CHARS``.
-    * Title is long and no description: split — the long string becomes
-      the description (preserved verbatim) and we derive a short label
-      for the title. This is the wall-of-text recovery path.
-    """
-    if not title:
-        return title, description
-    if len(title) <= LONG_TITLE_THRESHOLD:
-        return title, description
-    if description:
-        # Caller already split — respect their choice, just cap title so
-        # one bad input can't blow past the dashboard layout.
-        return title[:MAX_TITLE_CHARS], description
-    # Long title, no description: treat title as description, derive
-    # a short title from it.
-    short = _derive_short_title(title)
-    if not short:
-        # Pathological input (whitespace-only after splitting). Fall
-        # back to a hard cut so we always have *some* title.
-        short = title[:SHORT_TITLE_TARGET].rstrip() + "…"
-    return short, title
+# Title length policy lives in ``src.shared.task_titles`` so the agent
+# coordination tool can import it without crossing the trust boundary.
+# See ``_normalize_title_and_description`` (imported above) — applied in
+# ``Tasks.create`` as the authoritative server-side check.
 
 # Allowed status transitions. Keys are FROM, values are sets of valid TOs.
 # Terminal states (done / failed / cancelled) appear here with empty sets
